criando classe de serviço com dois métodos para interação com repositório

+++ src/main/java/dupradosantini/sostoolbackend/repositories/FilterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilterRepository extends JpaRepository<Filter, Integer> {
+public interface FilterRepository extends JpaRepository<Filter, Long> {
 }
